// src/models/interfaces/EmpresaRequest.ts
export interface EmpresaRequest {
  id: string;
  nome: string;
  nomeFantasia: string;
  endereco: string;
  numeroEndereco: number;
  cidade_id: string;
  numeroCnpj: string;
  numeroTelefone: string;
  complemento: string;
  bairro: string;
  email: string;
  nomeSecretaria: string;
  nomeDiretor: string;
  numeroCep: string;
  resolucao: string;
  declaracao: string;
  logoEmpresa: string;
}